// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ProductStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  ARCHIVED: 'ARCHIVED'
} as const

export type ProductStatus = (typeof ProductStatus)[keyof typeof ProductStatus]


export const OrderStatus = {
  PENDING_PAYMENT: 'PENDING_PAYMENT',
  PAID: 'PAID',
  PROCESSING: 'PROCESSING',
  READY_FOR_FULFILLMENT: 'READY_FOR_FULFILLMENT',
  SUPPLIER_ORDER_PENDING: 'SUPPLIER_ORDER_PENDING',
  SUPPLIER_ORDERED: 'SUPPLIER_ORDERED',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED',
  REFUNDED: 'REFUNDED',
  PAYMENT_FAILED: 'PAYMENT_FAILED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  AUTHORIZED: 'AUTHORIZED',
  PAID: 'PAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED',
  PARTIALLY_REFUNDED: 'PARTIALLY_REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const FulfillmentStatus = {
  PENDING: 'PENDING',
  ORDERED: 'ORDERED',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED'
} as const

export type FulfillmentStatus = (typeof FulfillmentStatus)[keyof typeof FulfillmentStatus]


export const CouponType = {
  PERCENT: 'PERCENT',
  FIXED: 'FIXED'
} as const

export type CouponType = (typeof CouponType)[keyof typeof CouponType]


export const AdminRole = {
  OWNER: 'OWNER',
  STAFF: 'STAFF'
} as const

export type AdminRole = (typeof AdminRole)[keyof typeof AdminRole]